requirements.txt
Django==4.2.4
django-safedelete==1.3.3
django-ordered-model==3.7.4
djangorestframework==3.14.0
nested-multipart-parser==1.5.0
django-restql==0.15.3
channels==4.0.0
django-timezone-field==5.0
Pillow==10.1.0
django-localflavor==4.0
Faker==0.7.0
factory-boy==3.3.0
django-queryable-properties==1.9.1
django-model-utils==4.3.1
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.0
Babel==2.14.0
django-simple-history==3.4.0
